# process images parallel while interacting with the browser