Little refactoring to process images parallel to browser calls. In my test case it was up to 36% faster, but that depends on the hardware, number of images to stitch together, browser, etc...

Stats:

- url: https://github.com/zinserjan/wdio-screenshot
- method: saveDocumentScreenshot
- Measured before and after makeAreaScreenshot with console.time

| | Before | After | Improvement |
| ------------- |:-------------:| -----:| -----:|
| iPad Air 2 iOS 10.2 landscape |  10825ms | 6888ms| 36% |
| Chrome v56 1024x768 | 6328ms | 5184ms | 18% |
| Firefox v51 | 1890ms | 1485ms | 21% |

*Note*: Firefox performs best here, cause it emits non-retina images and doesn't need the normalization -> less image processing.